Run necessary steps only

+++ scripts/seismic_autoencoding.py
@@ -25,13 +25,16 @@ print('')
 print('Dataloader creation...')
 device = 'cuda' if torch.cuda.is_available() else 'cpu'
 print('Device:', device)
-batch_size = 4
-dataloader = DataLoader(dataset,
-                        batch_size=batch_size,
-                        shuffle=True,
-                        collate_fn=dataloading.collate_fn)
-m, s = dataloading.compute_signal_mean_and_std(device, dataloader)
-dataloader.dataset.set_mean_and_std(m, s)
+if not os.path.isfile('data/dataloader.pth'):
+    batch_size = 4
+    dataloader = DataLoader(dataset,
+                            batch_size=batch_size,
+                            shuffle=True,
+                            collate_fn=dataloading.collate_fn)
+    m, s = dataloading.compute_signal_mean_and_std(device, dataloader)
+    dataloader.dataset.set_mean_and_std(m, s)
+    torch.save(dataloader, 'data/dataloader.pth')
+dataloader = torch.load('data/dataloader.pth')
 print('Dataloader saved')
 print('')
 
@@ -43,20 +46,23 @@ n_conv_channel_3 = 256
 lstm_hidden_size = 256
 n_lstm_layer = 1
 model = modeling.SignalAutoencoder(3, n_conv_channel_1, n_conv_channel_2,
-                                   n_conv_channel_3, lstm_hidden_size,
-                                   n_lstm_layer).to(device)
+                                n_conv_channel_3, lstm_hidden_size,
+                                n_lstm_layer).to(device)
 print(f'The model has {modeling.count_parameters(model)} parameters')
 loss_fn = modeling.sequence_l1
 optimizer = torch.optim.Adam(model.parameters(), lr=.001)
 print('')
 
-# Training (if no saved model)
+# Training
 print('Training...')
-model.train_one_epoch(device, dataloader, loss_fn, optimizer)
-torch.save(model.checkpoint, 'scripts/autoencoder.pt')
-plt.plot(model.checkpoint['loss_history'], label='Training')
+if not (os.path.isfile('data/dataloader.pth')
+        and os.path.isfile('scripts/autoencoder.pt')):
+    model.train_one_epoch(device, dataloader, loss_fn, optimizer)
+    torch.save(model.checkpoint, 'scripts/autoencoder.pt')
+checkpoint = torch.load('scripts/autoencoder.pt')
+plt.plot(checkpoint['loss_history'], label='Training')
 plt.xlabel('Iteration')
-plt.ylabel('Log-loss')
+plt.ylabel('loss')
 plt.legend()
 plt.savefig('viz/loss.png')
 print('Model saved')
@@ -64,7 +70,10 @@ print('')
 
 # Encoding
 print('Encoding...')
-model.load_state_dict(checkpoint['model_state_dict'])
-embeddings = model.encode(device, dataloader)
-torch.save(embeddings, 'data/embeddings.pt')
+if not (os.path.isfile('data/dataloader.pth')
+        and os.path.isfile('scripts/autoencoder.pt')
+        and os.path.isfile('scripts/embeddings.pt')):
+    model.load_state_dict(checkpoint['model_state_dict'])
+    embeddings = model.encode(device, dataloader)
+    torch.save(embeddings, 'data/embeddings.pt')
 print('Embeddings saved')
